Consolidate animations
